Adds an AstroCheck export
